working restructure
