+++ src/pages/Home.jsx
@@ -1,11 +1,11 @@
-import { Navbar }         from "@/components/Navbar";
-import { ThemeToggle }    from "@/components/ThemeToggle";
-import { StarBackground } from "@/components/StarBackground";
-import { HeroSection }    from "@/components/HeroSection";
-import { AboutSection }   from "@/components/AboutSection";
-import { SkillsSection }  from "@/components/SkillsSection";
-import { ProjectsSection } from "@/components/ProjectsSection";
-import { ContactSection } from "@/components/ContactSection";
+import { Navbar }         from "@/components/shared/Navbar";
+import { ThemeToggle }    from "@/components/shared/ThemeToggle";
+import { StarBackground } from "@/components/shared/StarBackground";
+import { HeroSection }    from "@/components/portfolio/HeroSection";
+import { AboutSection }   from "@/components/portfolio/AboutSection";
+import { SkillsSection }  from "@/components/portfolio/SkillsSection";
+import { ProjectsSection } from "@/components/portfolio/ProjectsSection";
+import { ContactSection } from "@/components/portfolio/ContactSection";
 import { Contact } from "lucide-react";
 
 export const Home = () => {
